factor: external-roleの処理をsync-externalに統合
